* changed planenum in side_t from int to uint16_t

+++ src/tools/ufo2map/map.h
@@ -59,7 +59,7 @@ typedef struct face_s {
 } face_t;
 
 typedef struct side_s {
-	int			planenum;
+	uint16_t	planenum;
 	int			texinfo;
 	winding_t	*winding;
 	struct side_s	*original;	/**< bspbrush_t sides will reference the mapbrush_t sides */
